Rename constants file to values
