+++ tests/common/test_exceptions.py
