+++ vigilant/common/values.py
